Read datasets.

+++ src/HDF5.NET/FileFormat/Level1/Level1A/BTree1Key/BTree1RawDataChunksKey.cs
@@ -6,14 +6,14 @@ namespace HDF5.NET
     {
         #region Constructors
 
-        public BTree1RawDataChunksKey(H5BinaryReader reader, int dimensionality)
+        public BTree1RawDataChunksKey(H5BinaryReader reader, byte dimensionality)
         {
             this.ChunkSize = reader.ReadUInt32();
             this.FilterMask = reader.ReadUInt32();
 
             this.ChunkOffsets = new ulong[dimensionality + 1];
 
-            for (int i = 0; i < dimensionality + 1; i++)
+            for (byte i = 0; i < dimensionality + 1; i++)
             {
                 this.ChunkOffsets[i] = reader.ReadUInt64();
             }
